Remove redundant below-merge check
