+++ crates/optimism/consensus/src/validation.rs
@@ -1,5 +1,4 @@
 use crate::proof::calculate_receipt_root_optimism;
-use alloc::vec::Vec;
 use alloy_consensus::TxReceipt;
 use alloy_primitives::{Bloom, B256};
 use reth_chainspec::{ChainSpec, EthereumHardforks};
